+++ src/main/java/com/bombombom/devs/user/service/UserService.java
@@ -1,8 +1,9 @@
-package com.bombombom.devs.user;
+package com.bombombom.devs.user.service;
 
-import com.bombombom.devs.user.controller.dto.SignupRequest;
+import com.bombombom.devs.user.UserRepository;
 import com.bombombom.devs.user.exception.ExistUsernameException;
 import com.bombombom.devs.user.models.User;
+import com.bombombom.devs.user.service.dto.SignupCommand;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -14,19 +15,15 @@ public class UserService {
     private final UserRepository userRepository;
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
-    public void addUser(SignupRequest signupRequest) {
-        if (userRepository.existsByUsername(signupRequest.username())) {
+    public void addUser(SignupCommand signupCommand) {
+        if (userRepository.existsByUsername(signupCommand.username())) {
             throw new ExistUsernameException();
         }
         User user = User.builder()
-            .username(signupRequest.username())
-            .password(bCryptPasswordEncoder.encode(signupRequest.password()))
-            .introduce(signupRequest.introduce())
-            .build();
+                .username(signupCommand.username())
+                .password(bCryptPasswordEncoder.encode(signupCommand.password()))
+                .introduce(signupCommand.introduce())
+                .build();
         userRepository.save(user);
-    }
-
-    public void removeUser(String username) {
-        userRepository.deleteByUsername(username);
     }
 }
