+++ src/main/java/com/bombombom/devs/user/service/dto/SignupCommand.java
@@ -1,0 +1,11 @@
+package com.bombombom.devs.user.service.dto;
+
+import lombok.Builder;
+
+@Builder
+public record SignupCommand(
+        String username,
+        String password,
+        String introduce
+) {
+}
